# api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: NurMed API
  version: v1
  description: API documentation for NurMed ERP/CRM backend
servers:
  - url: /api/v1
...
paths:
...
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token using refresh token cookie or X-Refresh-Token header
      operationId: refresh
      parameters:
        - in: header
          name: X-Refresh-Token
          required: false
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: Token refreshed
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Response'
                  - type: object
                    properties:
                      payload:
                        $ref: '#/components/schemas/AuthResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /auth/check-device:
    post:
      tags: [Auth]
      summary: Check device for verification
      operationId: checkDevice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CheckDeviceRequest'
      responses:
        '200':
          description: Device check response
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Response'
                  - type: object
                    properties:
                      payload:
                        $ref: '#/components/schemas/CheckDeviceResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /auth/verify-device:
    post:
      tags: [Auth]
      summary: Verify device with code
      operationId: verifyDevice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyDeviceRequest'
      responses:
        '200':
          description: Device verified
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Response'
                  - type: object
                    properties:
                      payload:
                        $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /users:
    get:
      tags: [Users]
      summary: List users
      operationId: getUsers
      security:
        - bearerAuth: []
      parameters:
        - in: header
          name: X-Company-ID
          required: false
          schema:
            type: integer
            format: int64
        - in: query
          name: company_id
          required: false
          schema:
            type: integer
            format: int64
        - in: query
          name: limit
          required: false
          schema:
            type: integer
        - in: query
          name: offset
          required: false
          schema:
            type: integer
      responses:
        '200':
          description: Users list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Response'
                  - type: object
                    properties:
                      payload:
                        type: array
                        items:
                          $ref: '#/components/schemas/UserResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    post:
      tags: [Users]
      summary: Create user with role scope
      operationId: createUser
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '200':
          description: User created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Response'
                  - type: object
                    properties:
                      payload:
                        $ref: '#/components/schemas/UserResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'

  # ── Admin / Companies ────────────────────────────────────────────────────────
...
  /admin/users:
    get:
      tags: [Admin / Users]
      summary: List users (admin)
      description: Доступно только суперадмину. Может фильтровать по любой компании.
      operationId: adminGetUsers
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: company_id
          required: false
          schema:
            type: integer
            format: int64
        - in: query
          name: status
          required: false
          schema:
            type: string
            enum: [active, blocked, invited, deleted]
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: Список пользователей
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Response'
                  - type: object
                    properties:
                      payload:
                        type: array
                        items:
                          $ref: '#/components/schemas/UserResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    post:
      tags: [Admin / Users]
      summary: Create user (admin)
      description: |
        Доступно только суперадмину. companyId передаётся в теле запроса.
        В отличие от POST /users — нет ограничений по scope, суперадмин может создать пользователя в любой компании.
      operationId: adminCreateUser
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '200':
          description: Пользователь создан
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Response'
                  - type: object
                    properties:
                      payload:
                        $ref: '#/components/schemas/UserResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Response'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Response'
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Response'
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Response'
...
  parameters:
...
    Limit:
      in: query
      name: limit
      required: false
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 200
    Offset:
      in: query
      name: offset
      required: false
      schema:
        type: integer
        default: 0
        minimum: 0
  schemas:
    Response:
      type: object
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: Успешно выполнено
        payload:
          nullable: true
...
    RefreshRequest:
      type: object
      properties:
        refreshToken:
          type: string
...
    CheckDeviceRequest:
      type: object
      required: [identifier]
      properties:
        identifier:
          type: string
          description: Username, email or phone number
    CheckDeviceResponse:
      type: object
      properties:
        deviceId:
          type: string
          description: Unique device identifier
        requiresVerification:
          type: boolean
          description: Whether verification code is required
        message:
          type: string
          description: Response message
        deviceInfo:
          $ref: '#/components/schemas/DeviceInfo'
    DeviceInfo:
      type: object
      properties:
        deviceId:
          type: string
          description: Unique device identifier
        deviceName:
          type: string
          description: Device name
        deviceType:
          type: string
          description: Device type (mobile, web, desktop)
        osType:
          type: string
          description: Operating system (iOS, Android, Windows, macOS, Linux)
        osVersion:
          type: string
          description: Operating system version
        appVersion:
          type: string
          description: Application version
        lastUsedAt:
          type: string
          format: date-time
          description: Last used timestamp
        isVerified:
          type: boolean
          description: Whether device is verified
        createdAt:
          type: string
          format: date-time
          description: Device creation timestamp
    VerifyDeviceRequest:
      type: object
      required: [identifier, verificationCode]
      properties:
        identifier:
          type: string
        verificationCode:
          type: string
    AuthResponse:
      type: object
      properties:
        accessToken:
          type: string
        accessTokenExpiresAt:
          type: string
          format: date-time
        tokenType:
          type: string
          example: Bearer
        userId:
          type: integer
          format: int64
        userName:
          type: string
        companyId:
          type: integer
          format: int64
          description: ID компании пользователя. 0 для суперадминов без привязки к компании.
...
    UserRoleAssignment:
      type: object
      required: [roleCode]
      properties:
        roleCode:
          type: string
          example: owner
        scopeType:
          type: string
          enum: [global, company, branch, warehouse]
          example: company
        scopeId:
          type: integer
          format: int64
        ownOnly:
          type: boolean
    CreateUserRequest:
      type: object
      required: [userName, password, firstName, role]
      properties:
        companyId:
          type: integer
          format: int64
        userName:
          type: string
        phone:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        status:
          type: string
          enum: [active, blocked, invited, deleted]
          example: active
        isSuperAdmin:
          type: boolean
        role:
          $ref: '#/components/schemas/UserRoleAssignment'
    UserResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        companyID:
          type: integer
          format: int64
        userName:
          type: string
        phone:
          type: string
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        status:
          type: string
        lastLoginAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...